[STEP 2] test: 로또 구매 개수를 구하는 기능에 대한 정상 및 예외 테스트 코드 작성
